requirements.txt
# --------- pytorch --------- #
torch==2.3.0
torchvision==0.18.0
lightning
torchmetrics>=0.11.4
datasets
transformers>=4.41.0
torch_geometric==2.3.0
--extra-index-url https://download.pytorch.org/whl/cu121
torch_scatter
torch_sparse
torch_cluster
tensorboard
-f https://data.pyg.org/whl/torch-2.3.0+cu121.html
scikit_learn
torchmetrics
jupyterlab
rootutils
pandas
# --------- hydra --------- #
hydra-core==1.3.2
hydra-colorlog==1.2.0
hydra-optuna-sweeper==1.2.0
# --------- loggers --------- #
# wandb
# neptune-client
# mlflow
# comet-ml
# aim>=3.16.2  # no lower than 3.16.2, see https://github.com/aimhubio/aim/issues/2550
# --------- others --------- #
rootutils       # standardizing the project root setup
pre-commit      # hooks for applying linters on commit
rich            # beautiful text formatting in terminal
pytest          # tests
#rdkit
biopython
scipy
git+https://github.com/ashtawy/data_engine.git